scoped: move code around, make `ExprTycker` smaller
